+++ Packages/KeyboardCore/Sources/KeyboardCore/KeyboardAppearanceSettings.swift
@@ -1,0 +1,3 @@
+public enum KeyboardAppearanceSettingsKey {
+    public static let liquidGlassMaterialEnabled = "keyboard_liquid_glass_material_enabled"
+}